+++ release_notes.md
@@ -1,8 +1,9 @@
-Fixed Issue #73. Please look at [ChangeLog](ChangeLog.md) for details on what has changed 
+This release contains a couple of security fixes. Please look at 
+[ChangeLog](ChangeLog.md) for details on what has changed 
 in the current version. 
 
 The binaries are cross-compiled with 
-[go-xbuild-go](https://github.com/muquit/go-xbuild-go.)
+[go-xbuild-go](https://github.com/muquit/go-xbuild-go)
 
 Do not forget to check checksums of the archives before using.